backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pydantic
pydantic-ai
fal-client
httpx
aiofiles
Pillow

